pre_arvoreb (sem testes)

+++ src/main/java/com/example/tpaedsiii/repository/bd/indexes/base/RegistroHash.java
@@ -5,7 +5,7 @@ public interface RegistroHash<T> {
 
   public int hashCode(); 
 
-  public int size(); 
+  public short size(); 
 
   public byte[] toByteArray() throws IOException;
